fix: 🐛 fix upload image route
